--- build.gradle
@@ -1,33 +1,0 @@
-plugins {
-    id 'java'
-}
-
-group 'com.sample.crawler'
-version '1.0-SNAPSHOT'
-
-sourceCompatibility = 1.8
-
-repositories {
-    mavenCentral()
-}
-
-dependencies {
-    // https://mvnrepository.com/artifact/org.jsoup/jsoup
-    compile group: 'org.jsoup', name: 'jsoup', version: '1.11.3'
-
-    testCompile group: 'junit', name: 'junit', version: '4.12'
-
-    // https://mvnrepository.com/artifact/org.slf4j/slf4j-api
-    compile group: 'org.slf4j', name: 'slf4j-api', version: '1.7.25'
-// https://mvnrepository.com/artifact/org.slf4j/slf4j-simple
-    compile group: 'org.slf4j', name: 'slf4j-simple', version: '1.7.25'
-// https://mvnrepository.com/artifact/org.slf4j/slf4j-log4j12
-    compile group: 'org.slf4j', name: 'slf4j-log4j12', version: '1.7.25'
-    compile 'org.slf4j:jcl-over-slf4j:1.7.25'
-    // https://mvnrepository.com/artifact/edu.uci.ics/crawler4j
-    compile group: 'edu.uci.ics', name: 'crawler4j', version: '4.1'
-// https://mvnrepository.com/artifact/com.google.guava/guava
-    compile group: 'com.google.guava', name: 'guava', version: '27.0.1-jre'
-
-    compileOnly 'org.projectlombok:lombok:1.18.4'
-}
