+++ common/build.gradle
@@ -1,0 +1,23 @@
+group = 'com.sample.common'
+version = '0.0.1-SNAPSHOT'
+
+
+apply plugin: 'java'
+
+repositories {
+    mavenCentral()
+}
+
+sourceCompatibility = '1.8'
+
+buildscript {
+    repositories {
+        mavenCentral()
+    }
+}
+
+dependencies {
+// https://mvnrepository.com/artifact/com.google.guava/guava
+    compile group: 'com.google.guava', name: 'guava', version: '27.0.1-jre'
+
+}
