+++ common/src/main/java/common/Hosting.java
