+++ common/src/main/java/common/Product.java
