+++ cralwer/src/main/java/com/sample/crawler/event/CrawlerEvent.java
@@ -1,0 +1,10 @@
+package com.sample.crawler.event;
+
+import lombok.AllArgsConstructor;
+import lombok.Getter;
+
+@Getter
+@AllArgsConstructor
+public class CrawlerEvent {
+  private String pipeId;
+}
