+++ scraping/build.gradle
@@ -14,7 +14,7 @@ apply plugin: 'java'
 apply plugin: 'org.springframework.boot'
 apply plugin: 'io.spring.dependency-management'
 
-group = 'com.sample'
+group = 'com.sample.scraping'
 version = '1-SNAPSHOT'
 sourceCompatibility = '1.8'
