--- src/main/java/makeshop/scraping/event/model/ScrapProductEvent.java
@@ -1,18 +1,0 @@
-package makeshop.scraping.event.model;
-
-import lombok.AllArgsConstructor;
-import lombok.Getter;
-import lombok.ToString;
-
-@Getter
-@ToString
-@AllArgsConstructor
-public class ScrapProductEvent {
-
-  private String domain;
-  private String shopId;
-  private String id;
-  private String link;
-  private String category;
-
-}
